requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
PyMySQL==1.1.0
stripe==7.0.0
